Integration helper cleanup: - move support code to 'support directory' - init page objects in helper
