+++ test/integration/support/integration-spec-helper.js
@@ -3,10 +3,12 @@ Zombie.site = 'http://localhost:3000';
 const browser = new Zombie();
 const screenshots = require('./screenshots');
 const knexCleaner = require('knex-cleaner');
-const knex = require('../../app/db').knex;
+const knex = require('../../../app/db').knex;
+const DashboardPage = require('../../common/page_objects/dashboard-page');
+const AddJobPage = require('../../common/page_objects/add-job-page');
 
 
-require('../../bin/www'); // This starts the web server, and ensures it is only
+require('../../../bin/www'); // This starts the web server, and ensures it is only
                           // started once. It is a misuse of "require", and
                           // should be improved.
 
@@ -21,4 +23,6 @@ module.exports = {
     return knexCleaner.clean(knex, { ignoreTables: ['knex_migrations'] });
   },
   browser,
+  dashboardPage: new DashboardPage(browser),
+  addJobPage: new AddJobPage(browser),
 };
